moving pptx parsing test into a test module
